only look for BIOS files in the libretro system directory (#122)
Don't look for BIOS files in the same directory as the executable,
instead obey the libretro system directory.

```cpp
 if (!g_retro_environment_callback(RETRO_ENVIRONMENT_GET_SYSTEM_DIRECTORY, &system_directory) || !system_directory)
   return GetProgramDirectoryRelativePath("system");
```
The code above reads "If we can't get the system directory from
libretro, return a directory named `system` located in the same
directory as the executable containing main()". This code is only used
to look for the path with the BIOS files.

On Unix, If RetroArch is /usr/bin/retroarch (or equivalent), it will try
to look for BIOS files at /usr/bin/system. On Android, non-root programs
will always be unable to look at that directory. Other platforms will
display similarly strange behavior.

This behavior feels more like an accident than an intentional design
choice. This commit removes it and all the code that implements it.